+++ app/controllers/publishers_controller.rb
@@ -169,93 +169,6 @@ class PublishersController < ApplicationController
     end
   end
 
-  # Records default currency preference
-  # If user does not have Uphold's `cards:write` scope, we redirect to Uphold to get authorization
-  # Card creation is done in #home
-  def confirm_default_currency
-    confirm_default_currency_params = publisher_confirm_default_currency_params
-    selected_currency = confirm_default_currency_params[:default_currency]
-
-    current_publisher.default_currency_confirmed_at = Time.now
-    current_publisher.save!
-
-    default_currency_changed = current_publisher.default_currency != selected_currency
-
-    if default_currency_changed
-      current_publisher.default_currency = selected_currency
-      current_publisher.save!
-
-      UploadDefaultCurrencyJob.perform_now(publisher_id: current_publisher.id)
-    end
-
-    # Check if the publisher is currently missing the ability to create cards
-    publisher_missing_write_scope = current_publisher.wallet.scope.exclude? "cards:write"
-
-    if publisher_missing_write_scope
-      # Redirect the publisher to Uphold in order to authorize card creation.
-      # Card will be created in #home when they return.
-      render(json: {
-        action: 'redirect',
-        status: t("publishers.confirm_default_currency_modal.redirecting"),
-        redirectURL: uphold_authorization_endpoint(current_publisher),
-        timeout: 3000,
-      }, status: 200)
-    else
-      create_uphold_card_for_default_currency_if_needed
-
-      render(json: {
-        action: 'refresh',
-        status: t("publishers.confirm_default_currency_modal.refreshing"),
-        timeout: 2000,
-      }, status: 200)
-    end
-  end
-
-  def uphold_verified
-    @publisher = current_publisher
-
-    # Ensure the uphold_state_token has been set. If not send back to try again
-    if @publisher.uphold_connection&.uphold_state_token.blank?
-      redirect_to(publisher_next_step_path(@publisher), alert: t(".uphold_error"))
-      return
-    end
-
-    # Catch uphold errors
-    uphold_error = params[:error]
-    if uphold_error.present?
-      redirect_to(publisher_next_step_path(@publisher), alert: t(".uphold_error"))
-      return
-    end
-
-    # Ensure the state token from Uphold matches the uphold_state_token last sent to uphold. If not send back to try again
-    state_token = params[:state]
-    if @publisher.uphold_connection&.uphold_state_token != state_token
-      redirect_to(publisher_next_step_path(@publisher), alert: t(".uphold_error"))
-      return
-    end
-
-    @publisher.uphold_connection.receive_uphold_code(params[:code])
-
-    begin
-      ExchangeUpholdCodeForAccessTokenJob.perform_now(publisher_id: @publisher.id)
-      @publisher.reload
-      @publisher.uphold_connection.reload
-    rescue Faraday::Error
-      Rails.logger.error("Unable to exchange Uphold access token with eyeshade")
-      redirect_to(publisher_next_step_path(@publisher), alert: t(".uphold_error"))
-      return
-    end
-
-    redirect_to(publisher_next_step_path(@publisher))
-  end
-
-  def disconnect_uphold
-    publisher = current_publisher
-    publisher.uphold_connection.disconnect_uphold
-    DisconnectUpholdJob.perform_later(publisher_id: publisher.id)
-
-    head :no_content
-  end
 
   def change_email
     @publisher = current_publisher
@@ -287,7 +200,8 @@ class PublishersController < ApplicationController
 
     @case = Case.find_by(publisher: current_publisher)
 
-    create_uphold_card_for_default_currency_if_needed
+    # TODO Add an if statement so it's more obvious
+    current_publisher.uphold_connection.create_uphold_card_for_default_currency
   end
 
   def statements
@@ -309,20 +223,6 @@ class PublishersController < ApplicationController
       statement_file_name = publishers_statement_file_name(@statement_period)
       statement_string = render_to_string :layout => "statement"
       send_data statement_string, filename: statement_file_name, type: "application/html"
-    end
-  end
-
-  def uphold_status
-    publisher = current_publisher
-    respond_to do |format|
-      format.json do
-        render(json: {
-          uphold_status: publisher.uphold_connection&.uphold_status.to_s,
-          uphold_status_summary: uphold_status_summary(publisher),
-          uphold_status_description: uphold_status_description(publisher),
-          uphold_status_class: uphold_status_class(publisher),
-        }, status: 200)
-      end
     end
   end
 
@@ -386,13 +286,6 @@ class PublishersController < ApplicationController
     end
   end
 
-  def create_uphold_card_for_default_currency_if_needed
-    if current_publisher.uphold_connection&.can_create_uphold_cards? &&
-      current_publisher.default_currency_confirmed_at.present? &&
-      current_publisher.wallet.address.blank?
-      CreateUpholdCardsJob.perform_now(publisher_id: current_publisher.id)
-    end
-  end
 
   def prepare_site_banner_data
     if current_publisher.default_site_banner_id.nil?
