--- app/jobs/upload_uphold_access_parameters_job.rb
@@ -1,17 +1,0 @@
-class UploadUpholdAccessParametersJob < ApplicationJob
-  queue_as :default
-
-  def perform(publisher_id:)
-    publisher = Publisher.find(publisher_id)
-    return if publisher.uphold_connection.uphold_verified
-    if publisher.uphold_connection.uphold_access_parameters.blank?
-      Rails.logger.info("Publisher #{publisher.id} is missing uphold_access_parameters. UpholdConnection #{publisher.uphold_connection.to_json}")
-      SlackMessenger.new(message: "🤔 Publisher #{publisher.id} is missing uphold_access_parameters.", channel: SlackMessenger::ALERTS).perform
-      return
-    end
-
-    PublisherWalletSetter.new(publisher: publisher).perform
-
-    publisher.uphold_connection.verify_uphold
-  end
-end
